CongressConnect/features/communication_outreach.py
```python
import smtplib
from email.mime.multipart import MIMEMultipart
from email.mime.text import MIMEText
from social_media_api import SocialMediaAPI

class CommunicationOutreach:

    def __init__(self, userSession, currentUser):
        self.userSession = userSession
        self.currentUser = currentUser
        self.socialMediaAPI = SocialMediaAPI()

    def send_newsletter(self, recipient_list, subject, message):
        # SMTP server configuration
        server = smtplib.SMTP('smtp.gmail.com', 587)
        server.starttls()

        # Login to the email account
        server.login(self.userSession['email'], self.userSession['password'])

        for recipient in recipient_list:
            # Create the email
            msg = MIMEMultipart()
            msg['From'] = self.userSession['email']
            msg['To'] = recipient
            msg['Subject'] = subject
            msg.attach(MIMEText(message, 'plain'))

            # Send the email
            server.send_message(msg)

        server.quit()

    def post_social_media_update(self, platform, message):
        self.socialMediaAPI.post_update(platform, message)

    def monitor_social_media(self, keywords):
        return self.socialMediaAPI.monitor_keywords(keywords)
```